Add another deploy workflow
We saw that we need more fine grained control over the files being
transmitted to S3. Especially setting things like cache time on a per
file basis, explicitly setting a charset and so on where used in some
projects.
We decided we want to keep the deploy mechanisms defined in a script in
respective projects. That way we also keep the ability to deploy a
project from the command line more easily (in case Github Actions are
not available).

I still think it makes sense to have a shared action that does all the
required setup and afterthoughts, like posting a notification on slack.